fix(web-worker-logic): remove circular dependency and consolidate
