+++ packages/web-worker-logic/index.ts
@@ -1,1 +1,3 @@
-export { gmailApiWorker, gmailBackgroundSyncWorker } from "./src/client";
+export { gmailApiWorker } from "./src/basic-client/gmail-api-worker";
+
+export { gmailBackgroundSyncWorker } from "./src/basic-client/gmail-bg-sync-worker";
