+++ packages/web-worker-logic/src/basic-client/gmail-bg-sync-worker.ts
@@ -1,0 +1,25 @@
+import { createTRPCClient } from "@trpc/client";
+import superjson from "superjson";
+
+import { workerLink } from "@skylar/trpc-web-workers";
+
+import type { GmailBackgroundSyncRouterType } from "../gmail-background-sync/router";
+import { loggerLinkConfig } from "./logger-config";
+
+export const gmailBackgroundSyncWorker = (emailAddress: string) =>
+  createTRPCClient<GmailBackgroundSyncRouterType>({
+    transformer: superjson,
+    links: [
+      loggerLinkConfig,
+      workerLink({
+        createWorker: () => {
+          return new SharedWorker(
+            new URL("../gmail-background-sync/worker.ts", import.meta.url),
+            {
+              name: `gmail-background-sync-worker-${emailAddress}`,
+            },
+          );
+        },
+      }),
+    ],
+  });
