+++ packages/web-worker-logic/src/gmail-api/lib/skylar-backend-connector.ts
@@ -1,15 +1,14 @@
-import { createTRPCClient, httpLink, loggerLink } from "@trpc/client";
+import { createTRPCClient, httpLink } from "@trpc/client";
 import superjson from "superjson";
 
 import type { AppRouter } from "@skylar/api";
 
+import { loggerLinkConfig } from "../../basic-client/logger-config";
+
 export const skylarBackendClient = createTRPCClient<AppRouter>({
   transformer: superjson,
   links: [
-    loggerLink({
-      enabled: (opts) =>
-        opts.direction === "down" && opts.result instanceof Error,
-    }),
+    loggerLinkConfig,
     httpLink({
       url: `${process.env.NEXT_PUBLIC_BACKEND_URL}/trpc`,
     }),
